Browser events moved out of prototype directory
