+++ modules/lo_event/examples/browser_events.html
@@ -10,9 +10,8 @@
 <p id="demo"></p>
 
 <div id="output"></div>
-<script src="../lo_event/prototype/lodash.min.js"></script>
 <script type="module">
-  import { subscribeToEvents } from '../lo_event/prototype/browser_events.js';
+  import { subscribeToEvents } from '../lo_event/browser_events.js';
 
   function logToDiv(eventType, event) {
       // Select the output div
